requirements.txt
aiohttp==3.9.1
aiohttp-cors==0.7.0
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
bidict==0.22.1
certifi==2023.11.17
charset-normalizer==3.3.2
clarifai==9.11.1
clarifai-grpc==9.11.5
contextlib2==21.6.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
imageio==2.33.1
ipython==8.20.0
jedi==0.19.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
ngrok==0.12.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
python-engineio==4.8.2
python-rapidjson==1.14
python-socketio==5.11.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
schema==0.7.5
simple-websocket==1.0.0
six==1.16.0
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.14.1
tritonclient==2.41.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
wsproto==1.2.0
yarl==1.9.4
